Dev

See merge request nvr/gcalendar_ruz!2

+++ gcalendar_ruz/main.py
@@ -1,103 +1,188 @@
-import time
 from datetime import datetime, timedelta
-import logging
-
+import asyncio
+from loguru import logger
+import time
 
 from core.apis.ruz_api import RuzApi
 from core.apis.calendar_api import GCalendar
 from core.db.models import Session, Room, OnlineRoom, Record, UserRecord, User
-from core.apis import nvr_api
-
-
-def create_logger(mode="INFO"):
-    logs = {"INFO": logging.INFO, "DEBUG": logging.DEBUG}
-
-    logger = logging.getLogger("ruz_logger")
-    logger.setLevel(logs[mode])
-
-    handler = logging.StreamHandler()
-    handler.setLevel(logs[mode])
-
-    formatter = logging.Formatter(
-        "%(levelname)-8s  %(asctime)s    %(message)s", datefmt="%d-%m-%Y %I:%M:%S %p"
-    )
-
-    handler.setFormatter(formatter)
-
-    logger.addHandler(handler)
-
-    return logger
-
-
-logger = create_logger()
+from core.apis.nvr_api import Nvr_Api
+from core.redis_caching.caching import redis_connect
 
 
 class CalendarManager:
     def __init__(self):
         self.session = Session()
         self.ruz_api = RuzApi()
-        self.calendar_api = GCalendar(
-            "/gcalendar_ruz/creds/creds.json",
-            "/gcalendar_ruz/creds/tokenCalendar.pickle",
-        )
+        self.nvr_api = Nvr_Api()
+        self.calendar_api = GCalendar()
 
     def __del__(self):
         self.session.close()
-        
-    def fetch_rooms(self):
-        ruz = self.session.query(OnlineRoom).filter_by(name="РУЗ").first()
-        jitsi = self.session.query(OnlineRoom).filter_by(name="Jitsi").first()
 
-        offline_rooms = [room.name for room in self.session.query(Room).all() if room.sources]
+    async def test_post_lesson(self, lesson: dict):
+        """ Post a lesson with empty event_id """
 
-        rooms = self.ruz_api.get_auditoriumoid()
+        lesson["gcalendar_event_id"] = ""
+        lesson["gcalendar_calendar_id"] = ""
 
-        for room in rooms:
-            try:
-                classes = self.ruz_api.get_classes(room["auditoriumOid"])
-                classes_len = len(classes)
-            except Exception as err:
-                logger.error(err, exc_info=True)
-                continue
+        data = await self.nvr_api.add_lesson(lesson)
 
-            for i in range(0, classes_len, 10):
-                chunk = classes[i : i + 10]
-                ruz_classes = [
-                    lesson
+        return data
+
+    async def post_lesson(self, lesson: dict, lesson_id: str, calendar_id: str):
+        """ Posts event to Google calendar and updates lesson in Erudite """
+
+        event = await self.calendar_api.create_event(ruz.calendar, lesson)
+        lesson["gcalendar_event_id"] = event["id"]
+        lesson["gcalendar_calendar_id"] = calendar_id
+        await self.nvr_api.update_lesson(lesson_id, lesson)
+
+        return event
+
+    async def add_lesson(self, lesson: dict, offline_rooms: list):
+        """ Adds lesson to Erudite and Google Calendar """
+
+        if lesson["ruz_url"] is None or "meet.miem.hse.ru" not in lesson["ruz_url"]:
+            logger.info("Adding ruz lesson")
+            data = await self.test_post_lesson(lesson)
+            code = data[0]
+            erudite_lesson = data[1]
+            if code == 201:
+                event = await self.post_lesson(
+                    lesson, erudite_lesson["id"], ruz.calendar
+                )
+                # time.sleep(0.6)
+
+            if lesson["ruz_auditorium"] in offline_rooms:
+                room = (
+                    self.session.query(Room)
+                    .filter_by(name=lesson["ruz_auditorium"])
+                    .first()
+                )
+                self.create_record(room, event)
+
+        elif lesson["ruz_url"] is not None and "meet.miem.hse.ru" in lesson["ruz_url"]:
+            logger.info("Adding jitsi lesson")
+            data = await self.test_post_lesson(lesson)
+            code = data[0]
+            erudite_lesson = data[1]
+            if code == 201:
+                event = await self.post_lesson(
+                    lesson, erudite_lesson["id"], jitsi.calendar
+                )
+
+    async def update_lesson(
+        self, lesson: dict, offline_rooms: list, lesson_id: str, event_id: str
+    ):
+        """ Updates lesson in Erudite and Google Calendar """
+
+        if lesson["ruz_url"] is None or "meet.miem.hse.ru" not in lesson["ruz_url"]:
+            logger.info("Updating ruz lesson")
+            event = await self.calendar_api.update_event(ruz.calendar, event_id, lesson)
+            lesson["gcalendar_event_id"] = event["id"]
+            lesson["gcalendar_calendar_id"] = ruz.calendar
+            await self.nvr_api.update_lesson(lesson_id, lesson)
+
+            if lesson["ruz_auditorium"] in offline_rooms:
+                room = (
+                    self.session.query(Room)
+                    .filter_by(name=lesson["ruz_auditorium"])
+                    .first()
+                )
+                self.create_record(room, event)
+
+        elif lesson["ruz_url"] is not None and "meet.miem.hse.ru" in lesson["ruz_url"]:
+            logger.info("Updating jitsi lesson")
+            event = await self.calendar_api.update_event(
+                jitsi.calendar, event_id, lesson
+            )
+            lesson["gcalendar_event_id"] = event["id"]
+            lesson["gcalendar_calendar_id"] = jitsi.calendar
+            await self.nvr_api.update_lesson(lesson_id, lesson)
+
+    async def synchronize_lesson(
+        self,
+        room_id: str,
+        lesson: dict,
+        offline_rooms: list,
+    ):
+        check_data = await self.nvr_api.check_lesson(lesson)
+        status = check_data[0]
+
+        # Lesson not found in Erudite, so we add it
+        if status == "Not found":
+            await self.add_lesson(lesson, offline_rooms)
+
+        # Lesson found in Erudite, but the data of this lesson has to be updated
+        elif status == "Update":
+            lesson_id = check_data[1]
+            event_id = check_data[2]
+            await self.update_lesson(lesson, offline_rooms, lesson_id, event_id)
+            # time.sleep(0.6)
+
+    async def synchronize_lessons_in_room(
+        self, room_id: str, offline_rooms: list, room_name: str
+    ):
+        lessons = await self.get_lessons_from_room(room_id)
+
+        if lessons:
+            logger.info(
+                f"""
+                Successfully got lessons for room {room_name}
+                Synchronizing lessons in calendar and Erudite
+                """
+            )
+            # Deletes lessons from Erudite if it doesn't exist in Ruz
+            await self.nvr_api.check_delete_Erudite_lessons(lessons, room_id)
+            time.sleep(0.2)
+
+            for i in range(0, len(lessons), 10):
+                chunk = lessons[i : i + 10]
+
+                tasks = [
+                    self.synchronize_lesson(room_id, lesson, offline_rooms)
                     for lesson in chunk
-                    if lesson["ruz_url"] is None
-                    or "meet.miem.hse.ru" not in lesson["ruz_url"]
                 ]
-                jitsi_classes = [
-                    class_
-                    for class_ in chunk
-                    if class_["ruz_url"] is not None
-                    and "meet.miem.hse.ru" in class_["ruz_url"]
-                ]
+                await asyncio.gather(*tasks)
 
-                logger.info(f"Adding ruz classes: {ruz_classes}")
-                for lesson in ruz_classes:
-                    event = self.calendar_api.create_event(ruz.calendar, lesson)
-                    lesson["gcalendar_event_id"] = event["id"]
-                    lesson["gcalendar_calendar_id"] = ruz.calendar
-                    nvr_api.add_lesson(lesson)
+    async def get_rooms(self):
+        offline_rooms = [room.name for room in self.session.query(Room).all()]
 
-                    if lesson['ruz_auditorium'] in offline_rooms:
-                        room = self.session.query(Room).filter_by(name=lesson['ruz_auditorium']).first()
-                        self.create_record(room, event)
+        rooms = await self.ruz_api.get_auditoriumoid()
 
-                logger.info(f"Adding jitsi classes: {jitsi_classes}")
-                for lesson in jitsi_classes:
-                    event = self.calendar_api.create_event(jitsi.calendar, lesson)
-                    lesson["gcalendar_event_id"] = event["id"]
-                    lesson["gcalendar_calendar_id"] = jitsi.calendar
-                    nvr_api.add_lesson(lesson)
+        tasks = [
+            self.synchronize_lessons_in_room(
+                room["auditoriumOid"], offline_rooms, room["number"]
+            )
+            for room in rooms
+        ]
 
-                time.sleep(10)
+        await asyncio.gather(*tasks)
 
         logger.info(
-            f"Creating events for {datetime.today().date() + timedelta(days=1)} done\n"
+            f"Created events for {datetime.today().date() + timedelta(days=1)} - {datetime.today().date() + timedelta(days=self.ruz_api.period)}"
         )
+
+    async def get_lessons_from_room(self, room_id: str) -> list:
+        """ Get lessons in room from ruz """
+
+        try:
+            lessons = await self.ruz_api.get_lessons(room_id)
+        except Exception as err:
+            lessons = None
+            logger.error(err)
+
+        return lessons
+
+    def get_online_rooms(self):
+        """ Get online rooms from DB """
+
+        global ruz
+        global jitsi
+
+        ruz = self.session.query(OnlineRoom).filter_by(name="РУЗ").first()
+        jitsi = self.session.query(OnlineRoom).filter_by(name="Jitsi").first()
 
     def create_record(self, room: Room, event: dict):
         start_date = event["start"]["dateTime"].split("T")[0]
@@ -122,19 +207,32 @@ class CalendarManager:
         self.session.add(user_record)
         self.session.commit()
 
-    def delete_online_events(self):
-        ruz = self.session.query(OnlineRoom).filter_by(name="РУЗ").first()
-        jitsi = self.session.query(OnlineRoom).filter_by(name="Jitsi").first()
+    async def delete_online_events(self):
+        while True:
+            events = await self.calendar_api.get_events(jitsi.calendar)
+            if len(events) == 0:
+                break
+            for event in events:
+                await self.calendar_api.delete_event(jitsi.calendar, event["id"])
 
-        events = self.calendar_api.get_events(jitsi.calendar)
-        for event in events:
-            self.calendar_api.delete_event(jitsi.calendar, event["id"])
+        while True:
+            events = await self.calendar_api.get_events(ruz.calendar)
+            if len(events) == 0:
+                break
+            for event in events:
+                await self.calendar_api.delete_event(ruz.calendar, event["id"])
 
-        events = self.calendar_api.get_events(ruz.calendar)
-        for event in events:
-            self.calendar_api.delete_event(ruz.calendar, event["id"])
+
+@logger.catch
+async def main():
+    await redis_connect()
+    manager = CalendarManager()
+
+    manager.get_online_rooms()
+
+    await asyncio.gather(manager.get_rooms())
 
 
 if __name__ == "__main__":
-    manager = CalendarManager()
-    manager.fetch_rooms()
+    loop = asyncio.get_event_loop()
+    loop.run_until_complete(main())
